Add an mu-plugin to set template and stylesheet on demand
This is a hack, but it works. To be used, this file must be copied or moved to the wp-content/mu-plugins directory so that it can alter the requested stylesheet and template as the page loads.

Unfortunately, these are set so early in the process that it's seemingly impossible to hook in any later.